backend/requirements.txt
fastapi
uvicorn
motor
pydantic-settings
python-jose[cryptography]
bcrypt==4.0.1
passlib
python-multipart
celery
redis
pymongo
